requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
